pyproject.toml
[tool.poetry]
name = "my_project"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^1.8.2"
python-docx = "^0.8.11"
PyMuPDF = "^1.18.19"
requests = "^2.26.0"
chromadb = "^0.5.5"
pdfplumber = "^0.5.28"
transformers = "^4.44.2"
torch = "^2.4.0"
sentence-transformers = "^3.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
